Back/requirements.txt
Django==4.1.1
django-cors-headers==3.13.0
django-model-utils==4.2.0
django-rest-swagger==2.2.0
django-swagger-ui==0.1.11
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
drf-yasg2==1.19.4
